Revert "Add nullable join"
This reverts commit 96bd4f1c
