+++ fennel/datasets/test_invalid_dataset.py
@@ -800,6 +800,41 @@ def test_dataset_incorrect_join():
         == "Cannot join with an intermediate dataset, i.e something defined inside a pipeline. Only joining against keyed datasets is permitted."
     )
 
+    with pytest.raises(TypeError) as e:
+
+        @dataset
+        class XYZ:
+            user_id: Optional[int]
+            agent_id: int
+            name: str
+            timestamp: datetime
+
+        @dataset(index=True)
+        class ABC:
+            user_id: int = field(key=True)
+            agent_id: int = field(key=True)
+            age: int
+            timestamp: datetime
+
+        @dataset
+        class XYZJoinedABC:
+            user_id: int
+            name: str
+            age: int
+            timestamp: datetime
+
+            @pipeline
+            @inputs(XYZ, ABC)
+            def create_pipeline(cls, a: Dataset, b: Dataset):
+                c = a.join(b, how="inner", on=["user_id", "agent_id"])  # type: ignore
+                return c
+
+    assert (
+        str(e.value)
+        == "Fields used in a join operator must not be optional in left schema, found `user_id` of "
+        "type `Optional[int]` in `'[Pipeline:create_pipeline]->join node'`"
+    )
+
 
 def test_dataset_incorrect_join_fields():
     with pytest.raises(ValueError) as e:
